backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Machine Learning
tensorflow==2.15.0
numpy<2.0,>=1.24.3  # TensorFlow 2.15.0 requires NumPy < 2.0
opencv-python==4.8.1.78
pillow==10.1.0

# Database - PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter for Python
alembic==1.13.0  # Database migrations (optional but recommended)

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Date and time
python-dateutil==2.8.2

# Utilities
aiosqlite==0.19.0  # Async SQLite support (backup option)